Notify on addition/removal failures
Determine which tags are:
a) Not able to be added due to either user or global searchtag restrictions; or
b) Not able to be removed due to the user not permitting artist set tags from being removed; while
i) Permitting otherwise valid tags to be added to the content item in question.

These tags will be passed to an error page to inform the user why their attempts to add/remove a tag seems to not be working.