Add 2FA logic path to signin controller
Determine if login.authenticate_bcrypt detected that 2FA was enabled from return tuple of userid/"2fa". If so, send user to 2fa token capture page. Userid is saved in the server-side session. Potential to-do: Figure out how to pass in the user's avatar for a "you are authenticating as Wesley, enter the 2FA token for that account?"