Reduce verbosity of 2FA instructions and increase readability
Improve the readability of the code strings (TOTP secret key and the recovery codes) by adding spaces when displaying them to the user. While the actual codes do not have spaces, and we explicitly tell the user that, be nice and make the entry fields 24 characters when it is appropriate to ingest a recovery code. The backend strips spaces, anyway.