Add function to validate claimed integers
While nothing is being passed directly through to the SQL-backend directly, Sentry is reporting ValueErrors when some functions straight pass user-input to int(). So handle it here in a sane way. If the input -is- an integer, then an integer will be returned; otherwise a zero is returned. Further input validation is left up to the function itself.

This, however, cuts down on many shenanigans or triggering of Sentry alerts.
[skip ci]